app/requirements.txt
Django==3.0.3
djangorestframework==3.11.0
requests==2.21.0
psycopg2==2.8.4
gunicorn==19.9.0